Feat: add ability to unselect starters in Tracker Library
I recently discovered Nomie while searching for habit tracking apps and
immediately fell in love with this one. Upon onboarding, however, I
accidentally selected a few too many starter trackers and was a slightly
irritated when clicking a starter a second time did not not remove it.

Thus, I decided to put my programming prowess to the test and implement
a little fix myself which is this commit.